--- a/Makefile
+++ b/Makefile
@@ -5,3 +5,6 @@ sample:
 
 sample-clean:
 	@docker compose -f docker-compose.sample.yaml down
+
+benchmark:
+	dotnet run -c Release --project ./tests/benchmark/Farfetch.LoadShedding.BenchmarkTests/Farfetch.LoadShedding.BenchmarkTests.csproj