--- a/src/Farfetch.LoadShedding/Events/Args/ItemDequeuedEventArgs.cs
+++ b/src/Farfetch.LoadShedding/Events/Args/ItemDequeuedEventArgs.cs
@@ -6,29 +6,17 @@ namespace Farfetch.LoadShedding.Events.Args
     /// <summary>
     /// Event args for the task dequeued event.
     /// </summary>
-    public class ItemDequeuedEventArgs : ItemEventArgs
+    public class ItemDequeuedEventArgs : TaskQueueEventArgs
     {
-        internal ItemDequeuedEventArgs(Priority priority, TimeSpan queueTime, int queueLimit, int queueCount)
-            : base(priority)
+        internal ItemDequeuedEventArgs(Priority priority, TimeSpan queueTime, IReadOnlyCounter queueCounter)
+            : base(priority, queueCounter)
         {
             this.QueueTime = queueTime;
-            this.QueueLimit = queueLimit;
-            this.QueueCount = queueCount;
         }
 
         /// <summary>
         /// Gets the time waiting in the queue.
         /// </summary>
         public TimeSpan QueueTime { get; }
-
-        /// <summary>
-        /// Gets the maximum number of items in the queue.
-        /// </summary>
-        public int QueueLimit { get; }
-
-        /// <summary>
-        /// Gets the current number of items in the queue.
-        /// </summary>
-        public int QueueCount { get; }
     }
 }