--- a/src/Farfetch.LoadShedding/Events/Args/ItemProcessedEventArgs.cs
+++ b/src/Farfetch.LoadShedding/Events/Args/ItemProcessedEventArgs.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using Farfetch.LoadShedding.Tasks;
 
 namespace Farfetch.LoadShedding.Events.Args
@@ -6,29 +6,17 @@ namespace Farfetch.LoadShedding.Events.Args
     /// <summary>
     /// Event args for task processed event.
     /// </summary>
-    public class ItemProcessedEventArgs : ItemEventArgs
+    public class ItemProcessedEventArgs : TaskItemEventArgs
     {
-        internal ItemProcessedEventArgs(Priority priority, TimeSpan processingTime, int concurrencyLimit, int concurrencyCount)
-            : base(priority)
+        internal ItemProcessedEventArgs(Priority priority, TimeSpan processingTime, IReadOnlyCounter concurrencyCounter)
+            : base(priority, concurrencyCounter)
         {
             this.ProcessingTime = processingTime;
-            this.ConcurrencyLimit = concurrencyLimit;
-            this.ConcurrencyCount = concurrencyCount;
         }
 
         /// <summary>
         /// Gets time spent to process the task.
         /// </summary>
         public TimeSpan ProcessingTime { get; }
-
-        /// <summary>
-        /// Gets the current concurrency limit.
-        /// </summary>
-        public int ConcurrencyLimit { get; }
-
-        /// <summary>
-        /// Gets the current concurrency items count.
-        /// </summary>
-        public int ConcurrencyCount { get; }
     }
 }