--- a/src/Farfetch.LoadShedding/Events/Args/ItemProcessingEventArgs.cs
+++ b/src/Farfetch.LoadShedding/Events/Args/ItemProcessingEventArgs.cs
@@ -1,27 +1,15 @@
-﻿using Farfetch.LoadShedding.Tasks;
+using Farfetch.LoadShedding.Tasks;
 
 namespace Farfetch.LoadShedding.Events.Args
 {
     /// <summary>
-    /// Event args for task procssing event.
+    /// Event args for task processing event.
     /// </summary>
-    public class ItemProcessingEventArgs : ItemEventArgs
+    public class ItemProcessingEventArgs : TaskItemEventArgs
     {
-        internal ItemProcessingEventArgs(Priority priority, int concurrencyLimit, int concurrencyCount)
-            : base(priority)
+        internal ItemProcessingEventArgs(Priority priority, IReadOnlyCounter concurrencyCounter)
+            : base(priority, concurrencyCounter)
         {
-            this.ConcurrencyLimit = concurrencyLimit;
-            this.ConcurrencyCount = concurrencyCount;
         }
-
-        /// <summary>
-        /// Gets the current concurrency limit.
-        /// </summary>
-        public int ConcurrencyLimit { get; }
-
-        /// <summary>
-        /// Gets the current concurrency items count.
-        /// </summary>
-        public int ConcurrencyCount { get; }
     }
 }