--- a/src/Farfetch.LoadShedding/Tasks/ConcurrentCounter.cs
+++ b/src/Farfetch.LoadShedding/Tasks/ConcurrentCounter.cs
@@ -1,4 +1,4 @@
-﻿namespace Farfetch.LoadShedding.Tasks
+namespace Farfetch.LoadShedding.Tasks
 {
     internal class ConcurrentCounter
     {
@@ -64,10 +64,7 @@
         {
             lock (this._locker)
             {
-                if (this._count > 0)
-                {
-                    this._count--;
-                }
+                this._count--;
 
                 return this._count;
             }