--- a/src/Farfetch.LoadShedding/Tasks/IReadOnlyCounter.cs
+++ b/src/Farfetch.LoadShedding/Tasks/IReadOnlyCounter.cs
@@ -1,0 +1,9 @@
+namespace Farfetch.LoadShedding.Tasks
+{
+    internal interface IReadOnlyCounter
+    {
+        int Count { get; }
+
+        int Limit { get; }
+    }
+}