--- a/tests/benchmark/Farfetch.LoadShedding.BenchmarkTests/Farfetch.LoadShedding.BenchmarkTests.csproj
+++ b/tests/benchmark/Farfetch.LoadShedding.BenchmarkTests/Farfetch.LoadShedding.BenchmarkTests.csproj
@@ -18,8 +18,4 @@
     <ProjectReference Include="..\..\..\src\Farfetch.LoadShedding\Farfetch.LoadShedding.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Compile Remove="$(MSBuildProjectDirectory)/../Traits.cs" />
-  </ItemGroup>
-
 </Project>