# Description

* Feature: Update the Grafana dashboard
  * Show queue time zero when we don't have anything in queue
  * Show concurrency and queue counters and limits on the same widget
* Fix: Queue was decreased twice when items were rejected
* Fix: Queue was not decreased if events are not defined
* Fix: Report to the events (Processing and Queue) the current counter values (count and limit) instead of those at the event time

Fixes #16 

## How Has This Been Tested?

Run the load tests and at the end of the tests, the queue and concurrency gauge metrics should report the value zero.

## Checklist

-   [x] My code follows the style guidelines of this project
-   [x] I have performed a self-review of my own code
-   [x] I have added tests to cover my changes
-   [x] I have made corresponding changes to the documentation

### Disclaimer

By sending us your contributions, you are agreeing that your contribution is made subject to the terms of our [Contributor Ownership Statement](https://github.com/Farfetch/.github/blob/master/COS.md)