@kikofps 2024-04-04
Hi @tiagodaraujo ,
Just releasing this pull request to a beta version, you can find [here](https://github.com/Farfetch/loadshedding/releases/tag/1.0.1-beta).